backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Background tasks
celery==5.3.6
flower==2.0.1

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0

# Logging and monitoring
python-json-logger==2.0.7
structlog==24.1.0

# Testing (optional for development)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Type checking
mypy==1.8.0
types-redis==4.6.0.20240106

# RAG Pipeline
sentence-transformers==2.3.1
faiss-cpu==1.7.4
tiktoken==0.5.2
numpy==1.24.3
langchain==0.1.5
langchain-community==0.0.16

